Do eating as a transaction (Bugfix required at rollback savepoint)

+++ src/main/java/wob/city/city/worker/FoodFactoryWorker.java
@@ -28,7 +28,7 @@ public class FoodFactoryWorker extends TimerTask {
     }
 
     private void generateRandomAmountOfFood(Food food) {
-        int randomAmount = Calculation.getRandomIntBetween(50000, 100000);
+        double randomAmount = Calculation.getRandomIntBetween(50000, 100000);
         foodFactoryDao.putOrAddFoodAmount(city.getName(), food.getName(), randomAmount);
     }
 }
